requirements.txt
streamlit>=1.37.0
pandas>=2.2.0
plotly>=5.22.0
reportlab>=4.2.0
python-dotenv>=1.0.1
openai>=1.40.0
numpy>=1.26.0
openpyxl>=3.1.0